Added basic shoulder solver
First adventure with custom resources in Godot. It was a bit finnicky with C# (GDScript worked fine), but I think I've got it down!

The solver itself doesn't really work, but I want the code infrastructure and basic algorithm to be present for the entire upper body, and then I'll go and make everything work as it should.